% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{add_db_filters}
\alias{add_db_filters}
\title{FUNCTION_TITLE}
\usage{
add_db_filters(filters, db_type, filters_first = FALSE)
}
\arguments{
\item{filters}{PARAM_DESCRIPTION}

\item{db_type}{PARAM_DESCRIPTION}

\item{filters_first}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
