% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{check_columns}
\alias{check_columns}
\title{Check column names for filters}
\usage{
check_columns(col_names, columns = NULL)
}
\arguments{
\item{col_names}{A vector of column names}

\item{columns}{A dataframe that contains equivalence between Gamebirds
columns and the package ones. Default: NULL}
}
\value{
A vector of same length as \emph{col_names} but containing only renamed
columns
}
\description{
When creating filters, either new columns names defined by the
\strong{gblincoln} package or column names provided by Gamebirds are accepted.
This function takes as input a vector of column names and replaces all
Gamebirds columns found by the package ones.
}
\seealso{
\code{\link{rename_columns}}
}
