% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{check_db_type}
\alias{check_db_type}
\title{Check database type}
\usage{
check_db_type(db_type)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
