% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{check_type}
\alias{check_type}
\title{FUNCTION_TITLE}
\usage{
check_type(type)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
