% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{clean_dataset}
\alias{clean_dataset}
\title{Clean a data file directly exported from Gamebirds}
\usage{
clean_dataset(
  df,
  rename_columns = TRUE,
  colnames = NULL,
  add_age_classes = TRUE,
  age_classes = NULL,
  add_sex_classes = TRUE,
  sex_classes = NULL,
  correct_recovery_years = TRUE
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{colnames}{A data frame with old column names (old_colnames) and
new column names (new_colnames), Default: NULL}

\item{recoveries}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
