% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{compare_harvest_rates}
\alias{compare_harvest_rates}
\title{FUNCTION_TITLE}
\usage{
compare_harvest_rates(
  df1 = NULL,
  df2 = NULL,
  filters1 = NULL,
  filters2 = NULL,
  plot = TRUE,
  check_overlap = TRUE,
  ...
)
}
\arguments{
\item{df1}{PARAM_DESCRIPTION, Default: NULL}

\item{df2}{PARAM_DESCRIPTION, Default: NULL}

\item{filters1}{PARAM_DESCRIPTION, Default: NULL}

\item{filters2}{PARAM_DESCRIPTION, Default: NULL}

\item{plot}{PARAM_DESCRIPTION, Default: TRUE}

\item{check_overlap}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
