% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{correct_recoveries_years}
\alias{correct_recoveries_years}
\title{FUNCTION_TITLE}
\usage{
correct_recoveries_years(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Correct the recovery year by considering all hunting done before
April to belong to the previous year
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
