% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{correct_recovery_years}
\alias{correct_recovery_years}
\title{Correct recovery years}
\usage{
correct_recovery_years(df)
}
\arguments{
\item{df}{The data frame to modify}
}
\value{
The modified dataframe
}
\description{
Correct the recovery year by considering all hunting done before
April to belong to the previous year.
}
\seealso{
\code{\link{clean_dataset}}
}
