% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamebirds}
\alias{gamebirds}
\title{Gamebirds Dataset}
\format{
A data frame with 16068 rows and 43 variables:
\describe{
  \item{\code{Add.Info}}{character COLUMN_DESCRIPTION}
  \item{\code{Age}}{character COLUMN_DESCRIPTION}
  \item{\code{B.Coordinate.Precision}}{double COLUMN_DESCRIPTION}
  \item{\code{B.Day.Code}}{character COLUMN_DESCRIPTION}
  \item{\code{B.Dir}}{double COLUMN_DESCRIPTION}
  \item{\code{B.Flyway}}{character COLUMN_DESCRIPTION}
  \item{\code{B.Lat}}{double COLUMN_DESCRIPTION}
  \item{\code{B.Long}}{double COLUMN_DESCRIPTION}
  \item{\code{B.Month}}{character COLUMN_DESCRIPTION}
  \item{\code{B.Region}}{character COLUMN_DESCRIPTION}
  \item{\code{B.Year}}{character COLUMN_DESCRIPTION}
  \item{\code{Band.Prefix.Plus}}{character COLUMN_DESCRIPTION}
  \item{\code{Band.Size}}{character COLUMN_DESCRIPTION}
  \item{\code{Band.Type}}{character COLUMN_DESCRIPTION}
  \item{\code{Count.of.Birds}}{double COLUMN_DESCRIPTION}
  \item{\code{Country_code}}{character COLUMN_DESCRIPTION}
  \item{\code{GISBLat}}{double COLUMN_DESCRIPTION}
  \item{\code{GISBLong}}{double COLUMN_DESCRIPTION}
  \item{\code{How.Aged}}{character COLUMN_DESCRIPTION}
  \item{\code{How.Sexed}}{character COLUMN_DESCRIPTION}
  \item{\code{Permit}}{character COLUMN_DESCRIPTION}
  \item{\code{Sex}}{character COLUMN_DESCRIPTION}
  \item{\code{Sp..Num.}}{character COLUMN_DESCRIPTION}
  \item{\code{State_code}}{character COLUMN_DESCRIPTION}
  \item{\code{Status}}{character COLUMN_DESCRIPTION}
  \item{\code{Age::VAGE}}{character COLUMN_DESCRIPTION}
  \item{\code{AI::VAI}}{character COLUMN_DESCRIPTION}
  \item{\code{BType::VBtype}}{character COLUMN_DESCRIPTION}
  \item{\code{BType::VText}}{character COLUMN_DESCRIPTION}
  \item{\code{coord_precision::LOCATION_ACCURACY_DESC}}{character COLUMN_DESCRIPTION}
  \item{\code{DayCode::Day.Span}}{character COLUMN_DESCRIPTION}
  \item{\code{How_aged::How.Aged.Description}}{character COLUMN_DESCRIPTION}
  \item{\code{How_sexed::How.Sexed.Description}}{character COLUMN_DESCRIPTION}
  \item{\code{Location_lu::COUNTRY_NAME}}{character COLUMN_DESCRIPTION}
  \item{\code{Location_lu::STATE_NAME}}{character COLUMN_DESCRIPTION}
  \item{\code{Month::VMonth}}{character COLUMN_DESCRIPTION}
  \item{\code{Permits::Permittee}}{character COLUMN_DESCRIPTION}
  \item{\code{Region::Flyway}}{character COLUMN_DESCRIPTION}
  \item{\code{Region::State}}{character COLUMN_DESCRIPTION}
  \item{\code{Sex::VSEX}}{character COLUMN_DESCRIPTION}
  \item{\code{Species.Game.Birds::SPEC}}{character COLUMN_DESCRIPTION}
  \item{\code{Species.Game.Birds::Species}}{character COLUMN_DESCRIPTION}
  \item{\code{Status::VStatus}}{character COLUMN_DESCRIPTION}
}
}
\usage{
gamebirds
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
