% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_ATBR_banding}
\alias{gb_ATBR_banding}
\title{Banding dataset for Atlantic Brants}
\format{
A data frame with 2072 rows and 50 variables:
}
\usage{
gb_ATBR_banding
}
\description{
Banding data extracted from Gamebirds for the Atlantic Brant.
The dataset has been loaded using the \code{\link{load_dataset}} function
}
\keyword{datasets}
