% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_ATBR_harvest}
\alias{gb_ATBR_harvest}
\title{US Harvest data for the Atlantic Brant}
\format{
A data frame with 20 rows and 5 variables:
}
\usage{
gb_ATBR_harvest
}
\description{
US Harvest data for the Atlantic Brant
}
\details{
DETAILS
}
\keyword{datasets}
