% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_ATBR_recoveries}
\alias{gb_ATBR_recoveries}
\title{Recovery dataset for Atlantic Brants}
\format{
A data frame with 14521 rows and 96 variables:
}
\usage{
gb_ATBR_recoveries
}
\description{
Recovery data extracted from Gamebirds for the Atlantic Brant.
The dataset has been loaded using the \code{\link{load_dataset}} function
}
\keyword{datasets}
