% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gb_banding}
\alias{gb_banding}
\title{Gamebirds Banding Dataset}
\format{
A data frame with 16068 rows and 43 variables:
\describe{
  \item{\code{add_info}}{integer COLUMN_DESCRIPTION}
  \item{\code{age_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{coordinate_precision}}{integer COLUMN_DESCRIPTION}
  \item{\code{day_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{dir}}{integer COLUMN_DESCRIPTION}
  \item{\code{flyway_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{lat}}{integer COLUMN_DESCRIPTION}
  \item{\code{long}}{integer COLUMN_DESCRIPTION}
  \item{\code{month_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{region_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{year}}{integer COLUMN_DESCRIPTION}
  \item{\code{band_prefix_plus}}{integer COLUMN_DESCRIPTION}
  \item{\code{band_size}}{character COLUMN_DESCRIPTION}
  \item{\code{band_type}}{character COLUMN_DESCRIPTION}
  \item{\code{count_of_birds}}{integer COLUMN_DESCRIPTION}
  \item{\code{country_code}}{character COLUMN_DESCRIPTION}
  \item{\code{GISBLat}}{double COLUMN_DESCRIPTION}
  \item{\code{GISBLong}}{double COLUMN_DESCRIPTION}
  \item{\code{how_aged}}{character COLUMN_DESCRIPTION}
  \item{\code{how_sexed}}{character COLUMN_DESCRIPTION}
  \item{\code{permit}}{integer COLUMN_DESCRIPTION}
  \item{\code{sex_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{sp_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{state_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{status_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{age}}{character COLUMN_DESCRIPTION}
  \item{\code{VAI}}{character COLUMN_DESCRIPTION}
  \item{\code{btype_vbtype}}{character COLUMN_DESCRIPTION}
  \item{\code{btype_text}}{character COLUMN_DESCRIPTION}
  \item{\code{location_accuracy_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{day_span}}{character COLUMN_DESCRIPTION}
  \item{\code{how_aged_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{how_sexed_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{country_name}}{character COLUMN_DESCRIPTION}
  \item{\code{state_name}}{character COLUMN_DESCRIPTION}
  \item{\code{month}}{character COLUMN_DESCRIPTION}
  \item{\code{permittee}}{character COLUMN_DESCRIPTION}
  \item{\code{flyway}}{character COLUMN_DESCRIPTION}
  \item{\code{state}}{character COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{SPEC}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{status}}{character COLUMN_DESCRIPTION}
  \item{\code{age_short}}{character COLUMN_DESCRIPTION}
}
}
\usage{
gb_banding
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
