% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_colnames}
\alias{gb_colnames}
\title{Mapping between Gamebirds column and gblincoln}
\format{
A data frame with 103 rows and 2 variables:
}
\usage{
gb_colnames
}
\description{
Default column renaming for the gblincoln package to ensure
consistency. old_columns are columns provided by Gamebirds and new_columns
the ones used in the gblincoln package.
}
\details{
DETAILS
}
\keyword{datasets}
