% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_reporting_probas}
\alias{gb_reporting_probas}
\title{DATASET_TITLE}
\format{
A data frame with 44 rows and 4 variables:
\describe{
  \item{\code{b.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{rho}}{double COLUMN_DESCRIPTION}
  \item{\code{var_rho}}{double COLUMN_DESCRIPTION}
  \item{\code{se_rho}}{double COLUMN_DESCRIPTION}
}
}
\usage{
gb_reporting_probas
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
