% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_all_direct_recoveries}
\alias{get_all_direct_recoveries}
\title{FUNCTION_TITLE}
\usage{
get_all_direct_recoveries(
  banding_filters,
  recoveries_filters,
  band_type = "all",
  banding_df = NULL,
  recoveries_df = NULL
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
