% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_db}
\alias{get_db}
\title{FUNCTION_TITLE}
\usage{
get_db(db_type)
}
\arguments{
\item{db_type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
