% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_direct_recoveries}
\alias{get_direct_recoveries}
\title{FUNCTION_TITLE}
\usage{
get_direct_recoveries(
  banding_df,
  recoveries_df,
  filters = NULL,
  banding_filters = NULL,
  recoveries_filters = NULL,
  filtered = FALSE
)
}
\arguments{
\item{banding_df}{PARAM_DESCRIPTION}

\item{recoveries_df}{PARAM_DESCRIPTION}

\item{filters}{PARAM_DESCRIPTION, Default: NULL}

\item{banding_filters}{PARAM_DESCRIPTION, Default: NULL}

\item{recoveries_filters}{PARAM_DESCRIPTION, Default: NULL}

\item{filtered}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
