% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_harvest_rate}
\alias{get_harvest_rate}
\title{Calculate the harvest rate based on reporting probabilities}
\usage{
get_harvest_rate(df = NULL, rho_df = NULL, ...)
}
\arguments{
\item{df}{A dataframe generated by  \code{\link{get_direct_recoveries}}, Default: NULL}

\item{rho_df}{A dataframe with reporting probabilities, Default: NULL}

\item{...}{If \code{df} is not provided, additionnal arguments passed to
\code{\link{get_direct_recoveries}} to generate the direct directories dataframe}
}
\value{
A dataframe with reporting probabilities and havest rate with
variance, standard error, confidence levels and coefficient of variation.
}
\description{
Takes a dataframe generated by \code{\link{get_direct_recoveries}},
a dataframe containing reproting probabilities and calculates the harvest rates
based on recovery data of banded birds. Variance, standard error, confidence
levels and coefficient of variation are also calculated for the harvest rate.
The harvest rate is estimated based on data described in Alisauskas \emph{et al.}, (2014)
}
\seealso{
\code{\link{get_direct_recoveries}}
}
