% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_lincoln_estimates}
\alias{get_lincoln_estimates}
\title{FUNCTION_TITLE}
\usage{
get_lincoln_estimates(
  df = NULL,
  harvest_df = NULL,
  harvest_correction_factor = 0.61,
  plot_estimates = TRUE,
  save_estimates = TRUE,
  save_path = ".",
  ...
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION, Default: NULL}

\item{harvest_df}{PARAM_DESCRIPTION, Default: NULL}

\item{harvest_correction_factor}{PARAM_DESCRIPTION, Default: 0.61}

\item{plot_estimates}{PARAM_DESCRIPTION, Default: TRUE}

\item{save_estimates}{PARAM_DESCRIPTION, Default: TRUE}

\item{save_path}{PARAM_DESCRIPTION, Default: '.'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
