% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_recoveries_summary}
\alias{get_recoveries_summary}
\title{FUNCTION_TITLE}
\usage{
get_recoveries_summary(df, by_band = FALSE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
