% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{get_species_data}
\alias{get_species_data}
\title{FUNCTION_TITLE}
\usage{
get_species_data(species_name, db_dtype = "banding", use_spec_code = TRUE)
}
\arguments{
\item{species_name}{PARAM_DESCRIPTION}

\item{db_dtype}{PARAM_DESCRIPTION, Default: 'banding'}

\item{use_spec_code}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
