% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lincoln_banding_filter}
\alias{lincoln_banding_filter}
\title{Performs filtering on the gb_banding dataset for Lincoln estimates}
\usage{
lincoln_banding_filter(filters = NULL, df = NULL)
}
\arguments{
\item{columns}{\link[gblincoln]{LINCOLN_BANDING_COLUMNS}, Default: LINCOLN_BANDING_COLUMNS}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
For time columns, i.e. 'year', 'month_code' and 'day_code',
it is possible to select a period by
providing a vector with two named values 'start' and 'end'
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
