% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{lincoln_filter_db}
\alias{lincoln_filter_db}
\title{Performs filtering on the gb_banding dataset for Lincoln estimates}
\usage{
lincoln_filter_db(
  filters = NULL,
  df = NULL,
  type = "banding",
  band_type = "all"
)
}
\arguments{
\item{columns}{\link[gblincoln]{LINCOLN_BANDING_COLUMNS}, Default: LINCOLN_BANDING_COLUMNS}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
For time columns, i.e. 'year', 'month_code' and 'day_code',
it is possible to select a period by
providing a vector with two named values 'start' and 'end'.
If a filter is NULL, it will be skipped
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
