% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns of a dataframe}
\usage{
rename_columns(df, columns)
}
\arguments{
\item{df}{The dataframe whose columns need to be renamed}

\item{columns}{A dataframe tha contains at least two columns:
`old_colnames` which contains the current column names found in df
`new_colnames` which contains the new column names}
}
\value{
A dataframe with renamed columns
}
\description{
This function renames the columns of a dataframe using another
dataframe with two columns `old_colnames` and `new_colnames`. Only the
columns found in `old_colnames` will be renamed
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
