% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{set_age_classes}
\alias{set_age_classes}
\title{Set age classes}
\usage{
set_age_classes(df, age_classes = NULL)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{age_classes}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
