% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{set_sex_classes}
\alias{set_sex_classes}
\title{FUNCTION_TITLE}
\usage{
set_sex_classes(df, sex_classes = NULL)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{sex_classes}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
