% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{summarize_bandings}
\alias{summarize_bandings}
\title{Summarize bandings}
\usage{
summarize_bandings(df)
}
\arguments{
\item{df}{A banding dataframe}
}
\value{
A summary dataframe with a new column named \emph{n_banded} which contains
the number of banded birds by year
}
\description{
Counts the number of bird banded by year. Grouping is done on the
\emph{b.year} column.
}
