% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{summarize_recoveries}
\alias{summarize_recoveries}
\title{Summarize recoveries}
\usage{
summarize_recoveries(df, by_band = FALSE)
}
\arguments{
\item{df}{A recovery dataframe.}

\item{by_band}{Should the results be sorted by band type? Default: FALSE}
}
\value{
The summarized dataframe
}
\description{
Counts the number of recoveries of birds banded in the same year.
If the data is not grouped by bands, it will be grouped by age class
(\emph{age_short} column), banding year(\emph{b.year}) and recovery year (\emph{r.corrected_year}).
}
