% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{FUNCTION_TITLE}
\usage{
update_list(new, old)
}
\arguments{
\item{new}{PARAM_DESCRIPTION}

\item{old}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
