% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{check_columns}
\alias{check_columns}
\title{FUNCTION_TITLE}
\usage{
check_columns(col_names)
}
\arguments{
\item{col_names}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
