% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{clean_dataset}
\alias{clean_dataset}
\title{Clean a data file directly exported from Gamebirds}
\usage{
clean_dataset(df, colnames = NULL, recoveries = FALSE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{colnames}{A data frame with old column names (old_colnames) and
new column names (new_colnames), Default: NULL}

\item{recoveries}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
