% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{compare_harvest_rates}
\alias{compare_harvest_rates}
\title{Compares two harvest rates dataframes created from different filters}
\usage{
compare_harvest_rates(
  df1 = NULL,
  df2 = NULL,
  filters1 = NULL,
  filters2 = NULL,
  plot = TRUE,
  check_overlap = TRUE,
  ...
)
}
\arguments{
\item{df1}{First dataframe to compare, Default: NULL}

\item{df2}{Second dataframe to compare, Default: NULL}

\item{filters1}{Filters used to create \emph{df1} if it is not provided, Default: NULL}

\item{filters2}{Filters used to create \emph{df2} if it is not provided, Default: NULL}

\item{plot}{Plot the comparisons, Default: TRUE}

\item{check_overlap}{Check if all confidence levels overlap, Default: TRUE}

\item{...}{Additional arguments needed by \code{\link{get_harvest_rate}} to
create the dataframes if they are not provided}
}
\value{
A logical specifiying if all confidence levels overlap if \emph{check_overlap}
is set to TRUE.
}
\description{
This function takes either two dataframes returned by
\code{\link{get_harvest_rate}} or creates two from different filters.
It then compares the harvest rates for the two dataframes over the years
to see if differences exist.
}
\seealso{
\code{\link{get_harvest_rate}}
}
