% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{DEFAULT_LINCOLN_FILTERS}
\alias{DEFAULT_LINCOLN_FILTERS}
\title{Default filters to prepare the gb_banding dataset for Lincoln estimates}
\format{
A list with column names as names and values for the desired filters
}
\usage{
DEFAULT_LINCOLN_FILTERS
}
\description{
This lists contains all the default filters used to preprare
gb_banding dataset to calculate Lincoln estimates.
add_info: 00 (normal), 01 (color band), 07 (double bands),
08 (temp marker-paint or dye), 25 (geolocators), 18 (blood sampled)
}
\keyword{datasets}
