% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{filter_database}
\alias{filter_database}
\title{Performs filtering on a dataset}
\usage{
filter_database(
  db,
  filters = NULL,
  columns = NULL,
  use_default_filters = TRUE,
  db_type = NULL,
  filters_first = FALSE
)
}
\arguments{
\item{db}{PARAM_DESCRIPTION}

\item{filters}{PARAM_DESCRIPTION, Default: NULL}

\item{columns}{PARAM_DESCRIPTION, Default: NULL}

\item{use_default_filters}{PARAM_DESCRIPTION, Default: TRUE}

\item{db_type}{PARAM_DESCRIPTION, Default: NULL}

\item{filters_first}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
For time columns, i.e. 'year', 'month_code' and 'day_code',
it is possible to select a period by
providing a vector with two named values 'start' and 'end'.
If a filter is NULL, it will be skipped
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
