% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_ATBR_harvest}
\alias{gb_ATBR_harvest}
\title{DATASET_TITLE}
\format{
A data frame with 20 rows and 5 variables:
\describe{
  \item{\code{b.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{harvest}}{integer COLUMN_DESCRIPTION}
  \item{\code{var_harvest}}{integer COLUMN_DESCRIPTION}
  \item{\code{se_harvest}}{integer COLUMN_DESCRIPTION}
  \item{\code{cl_harvest}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
gb_ATBR_harvest
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
