% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_ATBR_recoveries}
\alias{gb_ATBR_recoveries}
\title{DATASET_TITLE}
\format{
A data frame with 14521 rows and 96 variables:
\describe{
  \item{\code{GISRLong}}{double COLUMN_DESCRIPTION}
  \item{\code{GISRLat}}{double COLUMN_DESCRIPTION}
  \item{\code{e.object_name}}{character COLUMN_DESCRIPTION}
  \item{\code{MARPLOT_layer_name}}{character COLUMN_DESCRIPTION}
  \item{\code{MARPLOT_map_name}}{character COLUMN_DESCRIPTION}
  \item{\code{symbol}}{integer COLUMN_DESCRIPTION}
  \item{\code{color}}{character COLUMN_DESCRIPTION}
  \item{\code{idmarplot}}{integer COLUMN_DESCRIPTION}
  \item{\code{add_info}}{integer COLUMN_DESCRIPTION}
  \item{\code{age_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.coordinate_precision}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.day}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.dir}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.flyway_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.lat}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.long}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.month}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.region}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.country_code}}{character COLUMN_DESCRIPTION}
  \item{\code{b.state_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{band}}{integer COLUMN_DESCRIPTION}
  \item{\code{band_size}}{character COLUMN_DESCRIPTION}
  \item{\code{band_type_current}}{character COLUMN_DESCRIPTION}
  \item{\code{band_type_orig}}{character COLUMN_DESCRIPTION}
  \item{\code{cardinal_direction}}{character COLUMN_DESCRIPTION}
  \item{\code{distance}}{double COLUMN_DESCRIPTION}
  \item{\code{e.country_code}}{character COLUMN_DESCRIPTION}
  \item{\code{e.state_code}}{character COLUMN_DESCRIPTION}
  \item{\code{GISBLat}}{double COLUMN_DESCRIPTION}
  \item{\code{GISBLong}}{double COLUMN_DESCRIPTION}
  \item{\code{how_aged}}{character COLUMN_DESCRIPTION}
  \item{\code{how_obt}}{integer COLUMN_DESCRIPTION}
  \item{\code{how_sexed}}{character COLUMN_DESCRIPTION}
  \item{\code{hunt_season_surv}}{integer COLUMN_DESCRIPTION}
  \item{\code{b.marker_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{e.marker_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{e.min_age_at_enc}}{double COLUMN_DESCRIPTION}
  \item{\code{other_bands}}{character COLUMN_DESCRIPTION}
  \item{\code{permit}}{integer COLUMN_DESCRIPTION}
  \item{\code{pres_cond}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.coordinate_precision}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.create_date_month}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.create_date.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.day}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.dir}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.flyway_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.lat}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.long}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.month}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.region}}{integer COLUMN_DESCRIPTION}
  \item{\code{r.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{replaced_band_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{replaced_band_translated}}{character COLUMN_DESCRIPTION}
  \item{\code{same.block}}{integer COLUMN_DESCRIPTION}
  \item{\code{sex_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{sp_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{status_code}}{integer COLUMN_DESCRIPTION}
  \item{\code{who_reported}}{integer COLUMN_DESCRIPTION}
  \item{\code{why_pre_1994.report_method_after_1994}}{integer COLUMN_DESCRIPTION}
  \item{\code{age}}{character COLUMN_DESCRIPTION}
  \item{\code{VAI}}{character COLUMN_DESCRIPTION}
  \item{\code{b.location_accuracy_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{b.day_vrday}}{logical COLUMN_DESCRIPTION}
  \item{\code{b.month_name}}{character COLUMN_DESCRIPTION}
  \item{\code{b.state_name_short}}{character COLUMN_DESCRIPTION}
  \item{\code{b.state_name}}{character COLUMN_DESCRIPTION}
  \item{\code{btype_current_vbtype}}{character COLUMN_DESCRIPTION}
  \item{\code{btype_current_text}}{character COLUMN_DESCRIPTION}
  \item{\code{btype_vbtype}}{character COLUMN_DESCRIPTION}
  \item{\code{btype_text}}{character COLUMN_DESCRIPTION}
  \item{\code{condition_band_status}}{character COLUMN_DESCRIPTION}
  \item{\code{condition_condition}}{character COLUMN_DESCRIPTION}
  \item{\code{how_obt_text}}{character COLUMN_DESCRIPTION}
  \item{\code{how_aged_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{how_sexed_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{e.country_name}}{character COLUMN_DESCRIPTION}
  \item{\code{e.state_name}}{character COLUMN_DESCRIPTION}
  \item{\code{b.country_name}}{character COLUMN_DESCRIPTION}
  \item{\code{b.loc_state_name}}{character COLUMN_DESCRIPTION}
  \item{\code{permittee}}{character COLUMN_DESCRIPTION}
  \item{\code{r.location_accuracy_desc}}{character COLUMN_DESCRIPTION}
  \item{\code{r.day_vrday}}{character COLUMN_DESCRIPTION}
  \item{\code{b.flyway_name}}{character COLUMN_DESCRIPTION}
  \item{\code{rept_method_text}}{character COLUMN_DESCRIPTION}
  \item{\code{r.flyway_name}}{character COLUMN_DESCRIPTION}
  \item{\code{r.month_name}}{character COLUMN_DESCRIPTION}
  \item{\code{r.state_code}}{character COLUMN_DESCRIPTION}
  \item{\code{r.state_name}}{character COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{SPEC}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{status_text}}{character COLUMN_DESCRIPTION}
  \item{\code{who}}{character COLUMN_DESCRIPTION}
  \item{\code{age_short}}{character COLUMN_DESCRIPTION}
  \item{\code{r.corrected_year}}{double COLUMN_DESCRIPTION}
}
}
\usage{
gb_ATBR_recoveries
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
