% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_colnames}
\alias{gb_colnames}
\title{DATASET_TITLE}
\format{
A data frame with 103 rows and 2 variables:
\describe{
\item{\code{old_colnames}}{character COLUMN_DESCRIPTION}
\item{\code{new_colnames}}{character COLUMN_DESCRIPTION}
}
}
\usage{
gb_colnames
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
