% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{gb_reporting_probas}
\alias{gb_reporting_probas}
\title{Reporting probabilities for banding data}
\format{
An object of class \code{data.frame} with 44 rows and 4 columns.
}
\usage{
gb_reporting_probas
}
\description{
Reporting probabilities of banding data as described in
Arnold \emph{et al.} (2020)
}
\keyword{datasets}
