% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_db_type}
\alias{get_db_type}
\title{FUNCTION_TITLE}
\usage{
get_db_type(db_type, short = FALSE)
}
\arguments{
\item{db_type}{PARAM_DESCRIPTION}

\item{short}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
