% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_filters}
\alias{get_filters}
\title{FUNCTION_TITLE}
\usage{
get_filters(type, filters)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{filters}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
