% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_harvest_rate}
\alias{get_harvest_rate}
\title{FUNCTION_TITLE}
\usage{
get_harvest_rate(df = NULL, rho_df = NULL, ...)
}
\arguments{
\item{df}{PARAM_DESCRIPTION, Default: NULL}

\item{rho_df}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
