% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{get_lincoln_estimates}
\alias{get_lincoln_estimates}
\title{Get Lincoln estimates}
\usage{
get_lincoln_estimates(
  df = NULL,
  harvest_df = NULL,
  harvest_correction_factor = 0.61,
  plot_estimates = TRUE,
  save_estimates = TRUE,
  save_path = ".",
  ...
)
}
\arguments{
\item{df}{A harvest rate dataframe generated by \code{\link{get_harvest_rate}}
, Default: NULL}

\item{harvest_df}{A dataframe containing harvest data estimates for the target
species and target years. The dataframe must have a column named b.year,
Default: NULL}

\item{harvest_correction_factor}{A factor used to correct the harvest
estimations which tend to be overstimated. By default, uses the factor for
goose species collected after 1999 as described by Padding and Royle (2012),
Default: 0.61}

\item{plot_estimates}{Should the estimates be plotted, Default: TRUE}

\item{save_estimates}{Should the estimates be saved in csv format, Default: TRUE}

\item{save_path}{The path where estimates should be saved, Default: '.'}

\item{...}{Additional arguments needed by \code{\link{get_harvest_rate}} to
create the \emph{df} if it is not provided}
}
\value{
A dataframe containing the Lincoln estimates
}
\description{
This is the function that calculates the Lincoln estimates based
on the paper by Alisauskas \emph{et al.}, (2014) using harvest estimates and
a harvest rate dataframe created by \code{\link{get_harvest_rate}}
}
