% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_species_locations}
\alias{get_species_locations}
\alias{get_species_countries}
\alias{get_species_states}
\alias{get_species_flyways}
\title{List the locations available for a species}
\usage{
get_species_locations(
  df,
  species_code,
  type = c("country", "state", "flyway"),
  db_types = c("b", "r"),
  columns = NULL,
  return_names = TRUE,
  return_codes = TRUE,
  sort_by = NULL
)

get_species_countries(df, species_code, ...)

get_species_states(df, species_code, ...)

get_species_flyways(df, species_code, ...)
}
\arguments{
\item{df}{The dataframe in which to search}

\item{species_code}{The 4 letters short code for the desired species
(as found in the SPEC column of the gb_banding dataset)}

\item{type}{The type of data desired. Can be one of "country", "state" or
"flyway", Default: c("country", "state", "flyway")}

\item{columns}{The desired columns. By default returns all the location
related columns: "country_name", "state_name" and "flyway"}

\item{sort_by}{Column by which the results are sorted. Default: NULL, the
results are unsorted}

\item{db_type}{The type of data desired, banding or recovery. Can take any
value accepted by \code{\link{get_db_type}} but will only work for banding
or recovery data, Default: c("b", "r")}
}
\value{
If columns contains more than one element, a data frame with all
available combinations of the selected columns.
If columns contains only one element, a sorted vector of the unique values
of the selected column
}
\description{
This is a convenience function to list all available locations
for a given species. By default, a data frame is returned with available
countries, states and flyways
}
\details{
DETAILS
}
\section{Functions}{
\itemize{
\item \code{get_species_countries}: Convenience functions to list
all available countries for the given species.

\item \code{get_species_states}: Convenience functions to list
all available states for the given species.

\item \code{get_species_flyways}: Convenience functions to list
all available countries for the given species.
}}

\examples{
get_species_location("ATBR")

get_species_location("ATBR", "country_name")
# is equivalent to:
get_species_countries("ATBR")

}
