% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lincoln_banding_filter}
\alias{lincoln_banding_filter}
\title{FUNCTION_TITLE}
\usage{
lincoln_banding_filter(df = NULL, filters = NULL)
}
\arguments{
\item{columns}{\link[gblincoln]{LINCOLN_BANDING_COLUMNS}, Default: LINCOLN_BANDING_COLUMNS}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
