% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LINCOLN_FILTERS}
\alias{LINCOLN_FILTERS}
\title{gb_banding dataset columns for Lincoln estimates}
\format{
A character vector with 1 value
}
\usage{
LINCOLN_FILTERS
}
\description{
Columns selected for filtering the gb_banding dataset to calculate
Lincoln estimates
}
\keyword{datasets}
