% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_update}
\alias{list_update}
\title{Update a list with another one}
\usage{
list_update(old, new, new_first = FALSE)
}
\arguments{
\item{old}{A list to be updated}

\item{new}{A list with values to be added or updated}

\item{new_first}{Should the values found in the new list not present in the
old list be added at the beginning of the list or the end, Default: FALSE}
}
\value{
The updated list
}
\description{
This function takes two lists with named elements, the one to
be updated and the #' one with the updates. If elements from the new list are
not present in the one to be updated, they will be added. If they are presents,
their value will be changed. Note that if elements are not named, they will
not be updated.
}
