% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_update_bak}
\alias{list_update_bak}
\alias{list_update}
\title{FUNCTION_TITLE}
\usage{
list_update_bak(old, new, new_first = FALSE)

list_update(old, new, new_first = FALSE)
}
\arguments{
\item{old}{PARAM_DESCRIPTION}

\item{new}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
