% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{load_dataset}
\alias{load_dataset}
\title{FUNCTION_TITLE}
\usage{
load_dataset(
  path,
  rename_columns = TRUE,
  colnames = NULL,
  add_age_classes = TRUE,
  age_classes = NULL,
  add_sex_classes = TRUE,
  sex_classes = NULL,
  correct_recovery_years = TRUE,
  ...
)
}
\arguments{
\item{path}{PARAM_DESCRIPTION}

\item{rename_columns}{PARAM_DESCRIPTION, Default: TRUE}

\item{colnames}{PARAM_DESCRIPTION, Default: NULL}

\item{add_age_classes}{PARAM_DESCRIPTION, Default: TRUE}

\item{age_classes}{PARAM_DESCRIPTION, Default: NULL}

\item{add_sex_classes}{PARAM_DESCRIPTION, Default: TRUE}

\item{sex_classes}{PARAM_DESCRIPTION, Default: NULL}

\item{correct_recovery_years}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
