% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns of a dataframe}
\usage{
rename_columns(df, columns)
}
\arguments{
\item{df}{The dataframe whose columns need to be renamed}

\item{columns}{A dataframe tha contains at least two columns:
\code{old_colnames} which contains the current column names found in df
\code{new_colnames} which contains the new column names.}
}
\value{
A dataframe with renamed columns
}
\description{
This function renames the columns of a dataframe using another
reference dataframe with two columns \strong{old_colnames} and \strong{new_colnames}. Only the
columns found in \strong{old_colnames} will be renamed. By default, this package
ships with default values for renaming that #' can be found in
the \code{\link{gb_colnames}} object.
}
