% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{summarize_bandings}
\alias{summarize_bandings}
\title{Summarize bandings}
\usage{
summarize_bandings(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Counts the number of bird banded by year
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
