% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{summarize_recoveries}
\alias{summarize_recoveries}
\title{Summarize recoveries}
\usage{
summarize_recoveries(df, by_band = FALSE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{by_band}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Count the number of recoveries by year
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
