% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincoln.R
\name{summarize_recoveries}
\alias{summarize_recoveries}
\title{FUNCTION_TITLE}
\usage{
summarize_recoveries(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
